move api and form to blocks/utils + cleanup
